requirements.txt
numpy>=2.1
scipy>=1.14
PyYAML>=6.0
matplotlib>=3.9
Pillow>=11.0
mujoco>=3.7.0
open3d==0.19.0
torch>=2.11.0
